rework some code structure
