+++ src/main/java/frc/robot/Robot.java
@@ -8,6 +8,10 @@ import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 
+
+//change this to switch container
+import frc.robot.containers.prod.RobotContainer;
+
 /**
  * The {@link RobotContainer} holds all subsystems, commands, suppliers, etc. in
  * a command-based structure. The only action in {@link Robot} is to schedule
@@ -16,13 +20,13 @@ import edu.wpi.first.wpilibj2.command.CommandScheduler;
  * respective commands.
  */
 public class Robot extends TimedRobot {
-	private Command m_autonomousCommand;
+	private Command autonomousCommand;
 
-	private RobotContainer m_robotContainer;
+	public RobotContainer robotContainer;
 
 	@Override
 	public void robotInit() {
-		m_robotContainer = new RobotContainer();
+		robotContainer = new RobotContainer();
 	}
 
 	@Override
@@ -44,10 +48,10 @@ public class Robot extends TimedRobot {
 
 	@Override
 	public void autonomousInit() {
-		m_autonomousCommand = m_robotContainer.getAutonomousCommand();
+		autonomousCommand = robotContainer.getAutonomousCommand();
 
-		if (m_autonomousCommand != null) {
-			m_autonomousCommand.schedule();
+		if (autonomousCommand != null) {
+			autonomousCommand.schedule();
 		}
 	}
 
@@ -61,8 +65,8 @@ public class Robot extends TimedRobot {
 
 	@Override
 	public void teleopInit() {
-		if (m_autonomousCommand != null) {
-			m_autonomousCommand.cancel();
+		if (autonomousCommand != null) {
+			autonomousCommand.cancel();
 		}
 	}
