+++ src/main/java/frc/robot/constants/Controller.java
@@ -1,0 +1,9 @@
+package frc.robot.constants;
+
+public class Controller {
+    public static final int driveControllerPort = 0;
+    public static final double driveControllerJoystickDeadband = 0.06;
+
+    public static final int secondaryControllerPort = 1;
+    public static final double secondaryControllerJoystickDeadband = 0.1;
+}
