+++ src/main/java/frc/robot/constants/Gyro.java
@@ -1,0 +1,5 @@
+package frc.robot.constants;
+
+public class Gyro {
+    public static final int pigeonPort = 17;
+}
