+++ src/main/java/frc/robot/constants/VisionTargeter.java
@@ -1,0 +1,9 @@
+package frc.robot.constants;
+
+public class VisionTargeter {
+    public static final double kP = 0.1;
+    public static final double kI = 0.1;
+    public static final double kD = 0.1;
+
+    public static final double taMultiplier = 1.25;
+}
