+++ src/main/java/frc/robot/containers/prod/CommandCombinator.java
@@ -1,11 +1,11 @@
-package frc.robot;
+package frc.robot.containers.prod;
 
 import edu.wpi.first.math.VecBuilder;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.Subsystem;
-import frc.robot.RobotContainer.Subsystems;
+import frc.robot.containers.prod.RobotContainer.Subsystems;
 
 /**
  * Class for combining commands defined in {@link RobotContainer}. Efficent way
