+++ src/main/java/frc/robot/containers/prod/RobotContainer.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot;
+package frc.robot.containers.prod;
 
 import java.util.function.DoubleSupplier;
 import java.util.function.Supplier;
@@ -14,7 +14,9 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.RunCommand;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
-import frc.robot.RobotContainer.State.ControllerState;
+import frc.robot.Constants;
+import frc.robot.Robot;
+import frc.robot.containers.prod.RobotContainer.State.ControllerState;
 import frc.robot.subsystems.Controller;
 import frc.robot.subsystems.Pgyro;
 import frc.robot.subsystems.Swerve;
@@ -29,6 +31,8 @@ import io.github.oblarg.oblog.Logger;
  * commands.
  */
 public class RobotContainer {
+	public static final Constants constants = new Constants();
+
 	private SendableChooser<Command> autoChooser;
 
 	private class Controllers {
