--- src/main/java/frc/robot/data/GlobalState.java
@@ -1,5 +1,0 @@
-package frc.robot.data;
-
-public class GlobalState {
-    public static boolean isEnabled = false;
-}
